0.4.1: finish the grouping, type the protocol's numbers, and chain client options (#20)
* fix(core): decode screenshots returned as a data URI

GetSourceScreenshot returns "data:image/png;base64,..." rather than bare
Base64, so Convert.FromBase64String threw and the in-memory helper
returned null for every call. The file variant was unaffected because it
never touched the payload.

* test(example): assert values for the helpers added in v0.4

The suite covered the generated requests but not the hand written
conveniences, which is how the screenshot decode bug survived.

* docs: rewrite the README for the grouped surface

Every example still showed the flat extension methods, which are gone.

* feat(core)!: move event streams onto their category group

The 60 stream accessors were the last flat extensions on the client, so
half the surface was grouped and half was not. The protocol documents
requests and events under one set of category headings, and the groups
now hold both.

Renames the group types from {Category}RequestGroup to {Category}Group,
since they no longer hold only requests. Call sites go from
client.SceneCreatedStream() to client.Scenes.SceneCreatedStream().

* refactor(core): fold the client level helpers into one type

Both helper files were left as hollow banner comments when their methods
moved onto the category groups, and WaitForEventAsync had overloads split
across two static classes. The lambda CallBatchAsync overload also still
returned the untyped list while the builder overload returned
BatchResults, so which overload you picked silently decided whether you
got typed results.

* style: format the repository with csharpier

The tool was pinned in dotnet-tools.json but nothing ever ran it, so 69
of 99 files had drifted. Some of that drift was misleading rather than
cosmetic: the group structs kept the indentation of the extension blocks
their methods were lifted out of. CI now checks it.

* refactor(core): use the typed surfaces inside the helpers

Six helpers decided "not found" by substring matching the exception
message, so rewording a message would have turned a null return into a
throw. ObsWebSocketRequestException now exposes the reported status as
the protocol enum, and the helpers match on that.

SetInputMutesAsync still built its batch out of raw BatchRequestItem and
returned nothing, leaving per input failures visible only in the log. It
uses the typed builder now and returns the results.

FindSceneItemIdAsync returns int? so the Int32 suffixed duplicate is no
longer needed, and the Try prefixed forwarder announced for removal in
v0.4 is gone.

* docs: show the typed status filter in the errors section

* feat(core)!: map whole number protocol fields to int and long

The protocol has one numeric type because JSON has one, so every Number
became a double and callers read scene item ids, frame counts and byte
counts as floating point. Which fields are integral is not recoverable
from the definition: sceneItemId and inputVolumeMul are both Number with
a >= 0 restriction. The classification is therefore an explicit table,
not a rule over field names, because guessing wrong on a volume field
truncates it silently while an unlisted field only stays double.

32 field names become int or long, 10 stay double. A refresh that adds
an unlisted Number field reports OBSWSGEN012 rather than drifting in.

Verified against OBS 32.2.2 on both transports: MessagePack decodes the
retyped fields, which was the risk.

* feat(core)!: chain client options off the registration

WithAutoConnect hung off IServiceCollection, so it read as applying to
nothing and could not target one of two clients. AddObsWebSocketClient
now returns a builder, the shape AddHttpClient uses, and WithAutoConnect,
WithHealthCheck and WithReconnectPipeline chain off it. A named client
gets its own connection service, its own options instance and a health
check name that does not collide with the other client's.

Also fixes a test that drove a FakeTimeProvider with a real Task.Delay,
which cannot finish inside its own timeout when the suite runs in
parallel. It waits on a signal now.

* refactor(core)!: name the enum and the scene helpers for what they are

The protocol's RequestStatus enum shared a name with the RequestStatus
record carried on every response, which is the pair a caller has to
disambiguate to write a status filter. The enum is RequestStatusCode now.

SwitchSceneAsync took a switchToProgram flag and sat next to the named
shorthands, so the group appeared to offer five ways to switch a scene.
The flag versions are private, reached through SwitchProgramSceneAsync
and SwitchPreviewSceneAsync, and the missing preview-and-wait shorthand
now exists.

Two more connection tests drove real delays rather than the fake clock,
and a container resolution test carried a one second wall clock limit
with nothing in it that could hang. All three failed only under load.

* docs: bring the example and README onto the current surface

The example built a typed batch and then dropped it to Raw, and
registered the health check separately from the client. The README's
status filter still named the enum by its old name, and neither the
numeric typing nor the per client chain were documented.

+++ ObsWebSocket.Tests/ObsWebSocketClientIntegrationTests.cs
@@ -49,7 +49,9 @@ public class ObsWebSocketClientIntegrationTests
         _ = services.AddLogging(builder => builder.AddConsole().SetMinimumLevel(minLogLevel));
 
         // Bind the "ObsIntegration" section from configuration to the options class
-        _ = services.Configure<ObsIntegrationTestOptions>(configuration.GetSection("ObsIntegration"));
+        _ = services.Configure<ObsIntegrationTestOptions>(
+            configuration.GetSection("ObsIntegration")
+        );
 
         // Configure ObsWebSocketClientOptions based on ObsIntegrationTestOptions
         _ = services
@@ -112,15 +114,13 @@ public class ObsWebSocketClientIntegrationTests
         ILogger<ObsWebSocketClient> logger = s_serviceProvider.GetRequiredService<
             ILogger<ObsWebSocketClient>
         >();
-        IWebSocketMessageSerializer serializer = s_serviceProvider.GetRequiredService<
-            IWebSocketMessageSerializer
-        >();
+        IWebSocketMessageSerializer serializer =
+            s_serviceProvider.GetRequiredService<IWebSocketMessageSerializer>();
         IOptions<ObsWebSocketClientOptions> options = s_serviceProvider.GetRequiredService<
             IOptions<ObsWebSocketClientOptions>
         >();
-        IWebSocketConnectionFactory connectionFactory = s_serviceProvider.GetRequiredService<
-            IWebSocketConnectionFactory
-        >();
+        IWebSocketConnectionFactory connectionFactory =
+            s_serviceProvider.GetRequiredService<IWebSocketConnectionFactory>();
 
         return new ObsWebSocketClient(logger, serializer, options, connectionFactory);
     }
@@ -445,12 +445,15 @@ public class ObsWebSocketClientIntegrationTests
             idResponse,
             $"Could not get SceneItemId for '{s_testOptions.TestInputName}' in scene '{s_testOptions.TestSceneName}'. Ensure it exists."
         );
-        double sceneItemId = idResponse.SceneItemId;
+        int sceneItemId = idResponse.SceneItemId;
 
         // Get the transform
         GetSceneItemTransformResponseData? transformResponse =
             await client.SceneItems.GetSceneItemTransformAsync(
-                new GetSceneItemTransformRequestData(sceneItemId, sceneName: s_testOptions.TestSceneName!)
+                new GetSceneItemTransformRequestData(
+                    sceneItemId,
+                    sceneName: s_testOptions.TestSceneName!
+                )
             );
 
         Assert.IsNotNull(transformResponse, "GetSceneItemTransform response was null.");
@@ -526,7 +529,8 @@ public class ObsWebSocketClientIntegrationTests
         GetInputSettingsResponseData? response;
         try
         {
-            response = await client.Inputs.GetInputSettingsAsync(new GetInputSettingsRequestData(s_testOptions.TestInputName!)
+            response = await client.Inputs.GetInputSettingsAsync(
+                new GetInputSettingsRequestData(s_testOptions.TestInputName!)
             );
         }
         catch (ObsWebSocketException ex) when (ex.Message.Contains("ResourceNotFound"))
@@ -636,7 +640,8 @@ public class ObsWebSocketClientIntegrationTests
         await client.ConnectAsync(TestContext.CancellationToken);
         Assert.IsTrue(client.IsConnected);
 
-        GetSceneTransitionListResponseData? response = await client.Transitions.GetSceneTransitionListAsync();
+        GetSceneTransitionListResponseData? response =
+            await client.Transitions.GetSceneTransitionListAsync();
 
         Assert.IsNotNull(response, "GetSceneTransitionList response was null.");
         Assert.IsNotNull(response.Transitions, "Transitions list was null.");
@@ -680,7 +685,10 @@ public class ObsWebSocketClientIntegrationTests
 
         Assert.IsNotNull(response, "GetOutputList response was null.");
         Assert.IsNotNull(response.Outputs, "Outputs list was null.");
-        Assert.IsNotEmpty(response.Outputs, "Expected at least one output (e.g., Simple Output or Advanced).");
+        Assert.IsNotEmpty(
+            response.Outputs,
+            "Expected at least one output (e.g., Simple Output or Advanced)."
+        );
 
         OutputStub? firstOutput = response.Outputs.FirstOrDefault();
         Assert.IsNotNull(firstOutput, "First output stub was null.");
@@ -812,7 +820,8 @@ public class ObsWebSocketClientIntegrationTests
         GetInputDefaultSettingsResponseData? response;
         try
         {
-            response = await client.Inputs.GetInputDefaultSettingsAsync(new GetInputDefaultSettingsRequestData(inputKind)
+            response = await client.Inputs.GetInputDefaultSettingsAsync(
+                new GetInputDefaultSettingsRequestData(inputKind)
             );
         }
         catch (ObsWebSocketException ex) when (ex.Message.Contains("InvalidInputKind"))
@@ -851,10 +860,7 @@ public class ObsWebSocketClientIntegrationTests
         );
 
         // Assert some known default properties for text_gdiplus_v3 exist
-        Assert.IsTrue(
-            defaultSettingsDict.ContainsKey("font"),
-            "Expected 'font' default setting."
-        );
+        Assert.IsTrue(defaultSettingsDict.ContainsKey("font"), "Expected 'font' default setting.");
         Assert.AreEqual(
             JsonValueKind.Object,
             defaultSettingsDict["font"].ValueKind,
